x/hackatom/types/message_testmsg_test.go
package types

import (
	"testing"

	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
	"github.com/stretchr/testify/require"
	"github.com/tolik22869/hackatom/testutil/sample"
)

func TestMsgTestmsg_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgTestmsg
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgTestmsg{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgTestmsg{
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}
